Move pass contants/types into the new `pass` module.
